+++ fluidSimLib/gridDiffusionSolver3.h
@@ -17,8 +17,8 @@ public:
 
     virtual void solve(
         const scalarGrid3& source,
-        double diffusionCoefficient,
-        double timeIntervalInSeconds,
+        FloatType diffusionCoefficient,
+        FloatType timeIntervalInSeconds,
         scalarGrid3* dest,
         const scalarField3& boundarySdf = constantScalarField3(mathUtil::maxFloat()),
         const scalarField3& fluidSdf = constantScalarField3(-mathUtil::maxFloat())) = 0;
@@ -26,8 +26,8 @@ public:
 
     virtual void solve(
         const extendVectorGrid3& source,
-        double diffusionCoefficient,
-        double timeIntervalInSeconds,
+        FloatType diffusionCoefficient,
+        FloatType timeIntervalInSeconds,
         extendVectorGrid3* dest,
         const scalarField3& boundarySdf = constantScalarField3(mathUtil::maxFloat()),
         const scalarField3& fluidSdf = constantScalarField3(-mathUtil::maxFloat())) = 0;
@@ -35,8 +35,8 @@ public:
 
     virtual void solve(
         const faceCenteredGrid3& source,
-        double diffusionCoefficient,
-        double timeIntervalInSeconds,
+        FloatType diffusionCoefficient,
+        FloatType timeIntervalInSeconds,
         faceCenteredGrid3* dest,
         const scalarField3& boundarySdf = constantScalarField3(mathUtil::maxFloat()),
         const scalarField3& fluidSdf = constantScalarField3(-mathUtil::maxFloat())) = 0;
